Move chart-layout to internal
