+++ src/internal/components/cartesian-chart/styles.scss
@@ -113,3 +113,39 @@
 .filter-container {
   /* used in test-utils */
 }
+
+.chart-container {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+
+  &--fit-height {
+    height: 100%;
+  }
+}
+
+.chart-container__vertical {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
+
+.chart-container__horizontal {
+  display: flex;
+
+  &--fit-height {
+    flex: 1;
+  }
+}
+
+.container-chart-plot-fit-height-wrapper {
+  position: absolute;
+  inset: 0;
+}
+.container-bottom-labels-fit-height-wrapper {
+  position: absolute;
+  left: 0;
+  right: 0;
+  bottom: 0;
+}
